+++ src/components/util/published.jsx
@@ -4,8 +4,11 @@ import dayjs from "dayjs";
 import localeDa from "dayjs/locale/da";
 import localizedFormat from "dayjs/plugin/localizedFormat";
 import { useTranslation } from "react-i18next";
-import calculateIsPublished from "./helpers/calculate-is-published";
-
+import {
+  calculateIsPublished,
+  currentlyPlaying,
+  willPlayInTheFuture,
+} from "./helpers/published-helper";
 /**
  * @param {object} props The props.
  * @param {object} props.published Object with from and to.
@@ -14,17 +17,27 @@ import calculateIsPublished from "./helpers/calculate-is-published";
 function Published({ published }) {
   const { t } = useTranslation("common", { keyPrefix: "published" });
   const [isPublished, setIsPublished] = useState(false);
+  const [publishedStyleClass, setPublishedStyleClass] = useState("");
 
   useEffect(() => {
     dayjs.extend(localizedFormat);
   }, []);
 
   useEffect(() => {
-    setIsPublished(calculateIsPublished(published));
+    const isPlaylistPublished = calculateIsPublished(published);
+    setIsPublished(isPlaylistPublished);
+
+    if (!currentlyPlaying(published) && !willPlayInTheFuture(published)) {
+      setPublishedStyleClass("text-muted");
+    }
   }, [published]);
 
   if (!published.from && !published.to) {
-    return <div>{isPublished ? t("yes") : t("no")}</div>;
+    return (
+      <div className={publishedStyleClass}>
+        {isPublished ? t("yes") : t("no")}
+      </div>
+    );
   }
 
   let publishedFrom = "-";
@@ -33,15 +46,17 @@ function Published({ published }) {
   if (published.from) {
     publishedFrom = dayjs(published.from).locale(localeDa).format("LLLL");
   }
+
   if (published.to) {
     publishedTo = dayjs(published.to).locale(localeDa).format("LLLL");
   }
+
   return (
     <>
-      <div>
+      <div className={publishedStyleClass}>
         {t("from")}: {publishedFrom}
       </div>
-      <div>
+      <div className={publishedStyleClass}>
         {t("to")}: {publishedTo}
       </div>
     </>