+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- [#209](https://github.com/os2display/display-admin-client/pull/209)
+- change font on body tag
+- add playing component for list column 
+- restyle published column
+- fix proptypes errors in console from paginationbutton
+- speed up cypress in GA and add tests of playing-component
 
 - [#227](https://github.com/os2display/display-admin-client/pull/227)
   - Changed how theme is loaded following change to API